+++ src/logger.ts
@@ -1,13 +1,28 @@
 import * as log from 'loglevel';
+import prefix from 'loglevel-plugin-prefix';
+import chalk from 'chalk';
 
-const logger = log.getLogger('kana');
-let logLevel: log.LogLevelDesc = 'debug';
+prefix.reg(log);
+
+const colors: any = {
+  TRACE: chalk.magenta,
+  DEBUG: chalk.cyan,
+  INFO: chalk.blue,
+  WARN: chalk.yellow,
+  ERROR: chalk.red,
+};
+
+let logLevel: log.LogLevelDesc = 'debug'; // default
 if (process.env.LOG_LEVEL) {
   logLevel = process.env.LOG_LEVEL as log.LogLevelDesc;
 } else {
   // eslint-disable-next-line no-console
-  console.warn('Warning: Environment variable LOG_LEVEL is not set and has been defaulted to debug');
+  console.error(colors.WARN(`Warning: Environment variable LOG_LEVEL is not set and has been defaulted to ${logLevel}`));
 }
-logger.setLevel(logLevel);
 
-export default logger;
+export function getLogger(name: string) {
+  const logger = log.getLogger(name);
+  logger.setLevel(logLevel);
+  prefix.apply(logger, { format: (level, name, timestamp) => `[${timestamp}] ${colors[level.toUpperCase()](level)} ${name}:` });
+  return logger;
+}